PO-2087:
1. Changed 'version' type in legacy request and response from Integer to BigInteger
2. Renamed tests with better descriptions

+++ src/main/java/uk/gov/hmcts/opal/dto/legacy/AddPaymentTermsLegacyRequest.java
@@ -9,6 +9,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigInteger;
+
 @Data
 @Builder
 @NoArgsConstructor
@@ -30,7 +32,7 @@ public class AddPaymentTermsLegacyRequest {
 
     @JsonProperty("version")
     @NotNull
-    private Integer version;
+    private BigInteger version;
 
     @JsonProperty("payment_terms")
     @NotNull
